Background display and printf debug
In main, display a white background color with a some text on blue backcolor.
Following https://os.mbed.com/docs/mbed-os/v5.15/tutorials/debugging-using-printf-statements.html to have debug, mbed-os was retrograded to 5.15.2 instead of last version of master in order to get "Serial.h".
And "lib/mbed-os/components/**" is now considered (in .mbedignore) in order to build mbed-os with QSPI with this version.