Add i2c interface and get id of lsm6dsox
lsm6dsox is split in two parts: accelerometer and gyroscope.
Code is based on mbed version of lsm6dsox driver.
In order to check the correct link to component, id is asked.

Some conflicts appear with typedef that could be resolved with a common
class to both (e.g. component-lsm6dsox).
Some functions in driver can be explored next time.