Interrupt 2 and MLC documentation
Found a way to make INT2 work, it was routed on INT1 due to the UCF file config.
We have had some trouble finding out how the configuration registers act on reboot, sometimes they change values and sometimes they keep the same.
We are also facing a problem where the interrupt will be latched on the sensor, but the board won't get the notification, thus not reading nor clearing the interrupt.

We also worked on our LSM6DSOX documentation. We will try to complete the document with what we learn about the sensor and its MLC in particular. The structure isn't totally defined for now, we just tried to make it somehow readable.