Doxygen comments on MachineLearningCore class and interrupt debug and cleaning.
We commented most of the MachineLearningCore class methods with doxygen friendly syntax.

Started checking event queues and implementation of event handling, but no modifications to the code have been done yet.

Discovered that the MLC looses its trees upon power loss. We didn't know that and will need to consider adding the setup process somewhere in the init of the MLC.

We also reconfigured the MLC to use INT1 as it will eventually be the only pin available for interrupts.