First shot at data logging
A new folder has been added with the base of an app for data gathering
The goal of this app is to allow us to generate data sets for specific movements in order to use them to train the MLC

The idea would be to setup the app via serial from the PC, and then have a command to trigger the data collection

We still aren't sure if data should be sent as it is measured or on one big log at the end.

Have been struggling with serial and interrupts for the moment.