Working on data collection app
Interrupts did not work so we went for a polling method in order to make progress on the other functionnalities

We added a "basic" commands system that allow to change the configuration of the data logging via the serial terminal
These commands could be formalized if the app was interfaced with some computer app, but for now we are dealing with user input, so it has to be easy to use and failproof.

added a todo.md file used as a to-do list.