UCF processing
Two new applications to manage .ucf files

UCF_to_H_File_Conversion allows to take a .ucf and transform it into a working header file to insert in an mbed project in order to set up the MLC

UCF_Analysis aims to help us understand how the UCF works, we are using the register maps from the documentation to automatically comment the file.
The idea is to make the file much easier to read and to understand in order to pintpoint what can be understood with the given documentation and what is total mystery.