Implementing DataGatherer Class
We have been transforming the Data Gathering application in a class that handles all of it functionnalities in an easier way to set up

The idea was also to avoid using printf because it can be laggy and to allow communication over another serial port

A ticker is being implemented to print the data to serial at a given rate. We are still having trouble with the callback configuration.