Debugging DataGatherer
We have been trying to fix the calls to the DataGatherer callbacks "onSerial" and "onTick"

We implemented the queueing of the callbacks on the mbed event queue and we have increased the baud rate of the application.
So far everything seems to work if the data is sent in binary format.
We will try to update the data from the sensors via interrupt on the data_ready signal in order to lighten the printData method.