Working on embedded features implementation and documentation
We started working on the documentation for the embedded features ( activity, free fall, wake up, tilt , single/double tap, etc.).
We gathered all the informations we could in the todo.md file for now and will be moving them when it's completed.
The informations for now are mainly about the registers and bits impacting those embedded features.

On the features detection implementation we are still in the early stages and are still trying to figure out how to cleanly detect the features without interfering with others.