More work on embedded features tests and documentation
We have been working all day on the embedded features tests.
A lot of parameters are interconnected and registers are used for multiple things, thus making the overall understanding quite difficult.

So far we have been able to test
- free fall
- wake up
- sleep - motion/stationary
- 6D orientation

they seem to work fine, but oftentimes one config register corresponds to multiple features
So we are trying to clarify all the links in order to avoid modifying behaviour without being aware of that.