Add Pelion in Disco, configure mandatory
Pelion example is added in main, and reorganized to be done after
checking ethernet and qspi access.

Storage-size was mandatory in configuration for compilation and was added.

Now Disco is available in Pelion Portal with this configuration. The
counter is accessible and increment as Disco display it in Serial.

Update is not possible (yet)